Fix for the multiple worker setup
